Decoupled Label from Tag to support SKOS model
Decoupled label from Tag as there was a major design issue - label text had to be unique. This violated three basic requirements for SKOS scoping - the same text is legal for different languages, the same label is legal on different concepts and with multiple schemes you should expect the same label to reappear.
Tag equates to concept - in which case the tag ought to be the bound to the default namespace, which should represent a concept schema for tags, Maybe a different way of linking to Tag behaviour is required.